+++ integrations/consensus/src/main/java/org/pragmatica/consensus/net/netty/Handler.java
@@ -36,6 +36,13 @@ public class Handler extends SimpleChannelInboundHandler<Message.Wired> {
             if (msg instanceof Hello hello) {
                 helloHandler.accept(hello, ctx.channel());
             } else {
+                if (msg.getClass()
+                       .getSimpleName()
+                       .contains("Invoke")) {
+                    log.info("Handler routing InvokeMessage: {}",
+                             msg.getClass()
+                                .getName());
+                }
                 messageHandler.accept(msg);
             }
         } catch (Exception e) {
@@ -53,9 +60,9 @@ public class Handler extends SimpleChannelInboundHandler<Message.Wired> {
 
     @Override
     public void channelInactive(ChannelHandlerContext ctx) {
-        log.debug("Channel deactivated: {}",
-                  ctx.channel()
-                     .remoteAddress());
+        log.info("Channel deactivated: {}",
+                 ctx.channel()
+                    .remoteAddress());
         peerDisconnected.accept(ctx.channel());
     }